[HOP] improve naming for subgraph inputs
Previously, subgraph input names were whatever the input proxies were,
which were confusing. This PR changes those names to be
whatever the names of the arguments the functions being
speculate_subgraph'ed are. This is best-effort: if we can't figure it
out then we go back to the previous strategy.

Test Plan:
- existing expecttests

ghstack-source-id: 70698f509eb0ba2d91a83aef1f7e63b8569fd68b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130255